// Angular_project/frontend/src/app/models/user.ts
export class User{
    firstname: string;
    lastname: string;
    username: string;
    password: string;
    type: string;
    adress: string;
    phone: string;
    mail: string;
    licenceNumber: number;
    specialization: string;
    branch: string;
    image: string;
    added: boolean;
}